Adjust player animation so collider changes based on animation.
When the player changes directions, the collider size also changes so it better adheres to the player's character.
Map size and number of obstacles are now fully randomized and working from the main menu.
If pressing play from the Game scene, a default map size of 50 and default of 40 obstacles will spawn.