Fix lives issue WOOOOOO!
The issue was caused from the bonus lives the player was given. The bonus lives were granted if the current score was divided by 100 and the remainder equaled 0. Since you can have 0 score, divide it by 100 and end up with 0 score, the lives would be increased unless you gained some score.
To resolve this issue, there is now another condition that checks if the player's score is above 0. If it is, then the division occurs.