Fix player errors related to visibility and interactibility.
Player could not be seen after taking damage. This was because classColor was not set, so it is now set to a feault white color if it is not given a specific color.
The other issue as that the player's collider was marked as trigger, so this has been unchecked to allow for collision with obstacles and enemies.